Code smell fix
