+++ package.json
@@ -49,8 +49,8 @@
     "scripts": {
         "lint": "eslint src/**/*.tsx",
         "lint-fix": "eslint --fix src/**/*.tsx",
-        "start": "react-app-rewired start",
-        "build": "react-app-rewired build",
+        "start": "react-app-rewired --openssl-legacy-provider start",
+        "build": "react-app-rewired build --openssl-legacy-provider build",
         "test": "react-app-rewired test",
         "test-coverage:watch": "npm run test -- --coverage",
         "test-coverage": "npm run test -- --coverage --watchAll=false",
