+++ src/setupProxy.js
@@ -4,7 +4,7 @@ module.exports = function (app) {
     app.use(
         '/orchestrator',
         createProxyMiddleware({
-            target: 'http://demo.devtron.info:32080/',
+            target: 'http://20.241.140.189:30690/',
             changeOrigin: true,
             logLevel: 'info',
             secure: false,
